Moved starter to devoptions
