+++ app/src/main/java/org/oppia/android/app/application/ApplicationComponent.kt
@@ -5,8 +5,8 @@ import androidx.work.Configuration
 import dagger.BindsInstance
 import dagger.Component
 import org.oppia.android.app.activity.ActivityComponent
-import org.oppia.android.app.drawer.DeveloperOptionsModule
-import org.oppia.android.app.drawer.DeveloperOptionsStarterModule
+import org.oppia.android.app.devoptions.DeveloperOptionsModule
+import org.oppia.android.app.devoptions.DeveloperOptionsStarterModule
 import org.oppia.android.app.player.state.hintsandsolution.HintsAndSolutionConfigModule
 import org.oppia.android.app.shim.IntentFactoryShimModule
 import org.oppia.android.app.shim.ViewBindingShimModule
