+++ app/src/main/java/org/oppia/android/app/devoptions/DeveloperOptionsStarterImpl.kt
@@ -1,8 +1,7 @@
-package org.oppia.android.app.drawer
+package org.oppia.android.app.devoptions
 
 import android.content.Context
 import android.content.Intent
-import org.oppia.android.app.devoptions.DeveloperOptionsActivity
 import javax.inject.Inject
 
 /** Binds implementation of DeveloperOptionsStarter. */
